feat: personal review first pass; add func to validate custom config; fixed potential panic when decoding payload from redis; fixed potential divide by zero error; delete stale keys after a while; added a func to get breaker error; added exit case in main-loop; set expiry on all breaker keys;

+++ config/config.go
@@ -81,7 +81,7 @@ var DefaultConfiguration = Configuration{
 		FailureCount:                10,
 		SuccessThreshold:            5,
 		ObservabilityWindow:         5,
-		NotificationThresholds:      []int{5, 10},
+		NotificationThresholds:      []uint64{5, 10},
 		ConsecutiveFailureThreshold: 10,
 	},
 	Auth: AuthConfiguration{
@@ -272,14 +272,14 @@ type RetentionPolicyConfiguration struct {
 }
 
 type CircuitBreakerConfiguration struct {
-	SampleRate                  int     `json:"sample_rate" envconfig:"CONVOY_CIRCUIT_BREAKER_SAMPLE_RATE"`
-	FailureCount                int     `json:"failure_count" envconfig:"CONVOY_CIRCUIT_BREAKER_ERROR_COUNT"`
-	ErrorTimeout                int     `json:"error_timeout" envconfig:"CONVOY_CIRCUIT_BREAKER_ERROR_TIMEOUT"`
-	FailureThreshold            float64 `json:"failure_threshold" envconfig:"CONVOY_CIRCUIT_BREAKER_FAILURE_THRESHOLD"`
-	SuccessThreshold            int     `json:"success_threshold" envconfig:"CONVOY_CIRCUIT_BREAKER_SUCCESS_THRESHOLD"`
-	ObservabilityWindow         int     `json:"observability_window" envconfig:"CONVOY_CIRCUIT_BREAKER_OBSERVABILITY_WINDOW"`
-	NotificationThresholds      []int   `json:"notification_thresholds" envconfig:"CONVOY_CIRCUIT_BREAKER_NOTIFICATION_THRESHOLDS"`
-	ConsecutiveFailureThreshold int     `json:"consecutive_failure_threshold" envconfig:"CONVOY_CIRCUIT_BREAKER_CONSECUTIVE_FAILURE_THRESHOLD"`
+	SampleRate                  uint64   `json:"sample_rate" envconfig:"CONVOY_CIRCUIT_BREAKER_SAMPLE_RATE"`
+	FailureCount                uint64   `json:"failure_count" envconfig:"CONVOY_CIRCUIT_BREAKER_ERROR_COUNT"`
+	ErrorTimeout                uint64   `json:"error_timeout" envconfig:"CONVOY_CIRCUIT_BREAKER_ERROR_TIMEOUT"`
+	FailureThreshold            float64  `json:"failure_threshold" envconfig:"CONVOY_CIRCUIT_BREAKER_FAILURE_THRESHOLD"`
+	SuccessThreshold            uint64   `json:"success_threshold" envconfig:"CONVOY_CIRCUIT_BREAKER_SUCCESS_THRESHOLD"`
+	ObservabilityWindow         uint64   `json:"observability_window" envconfig:"CONVOY_CIRCUIT_BREAKER_OBSERVABILITY_WINDOW"`
+	NotificationThresholds      []uint64 `json:"notification_thresholds" envconfig:"CONVOY_CIRCUIT_BREAKER_NOTIFICATION_THRESHOLDS"`
+	ConsecutiveFailureThreshold uint64   `json:"consecutive_failure_threshold" envconfig:"CONVOY_CIRCUIT_BREAKER_CONSECUTIVE_FAILURE_THRESHOLD"`
 }
 
 type AnalyticsConfiguration struct {
